Commit book list page

+++ ASP.Server/Data/Book.cs
@@ -20,7 +20,7 @@ namespace ASP.Server.Model
 
         public String Auteur { get; set; }
 
-        public int Prix { get; set; }
+        public Double Prix { get; set; }
       
         public virtual ICollection<Genre> Genres { get; set; }
